# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Elgiganten Repair API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/"
        }
    ],
    "paths": {
...
        "/afhentningstyper": {
            "get": {
                "tags": [
                    "Sager - Afhentnings Typer"
                ],
                "summary": "Liste af Afhentnings Typer",
                "description": "",
                "operationId": "afhentningstypesList",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "navn": {
                                            "type": "string",
                                            "example": "Bedste Afhentnings Type Navn"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 02:28:47"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 03:15:46"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "Navn": "Bedste Afhentnings Type Navn",
                                        "created_at": "31-03-2022 02:28:47",
                                        "updated_at": "31-03-2022 03:15:46"
                                    },
                                    {
                                        "id": 1,
                                        "Navn": "Bedste Afhentnings Type Navn",
                                        "created_at": "31-03-2022 02:28:47",
                                        "updated_at": "31-03-2022 03:15:46"
                                    }
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Ingen Afhentnings Typer"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sager - Afhentnings Typer"
                ],
                "summary": "Opret en ny Afhentnings Type",
                "description": "",
                "operationId": "afhentningstypesCreate",
                "parameters": [
                    {
                        "name": "navn",
                        "in": "query",
                        "description": "Afhentnings Type navnet",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Afhentnings Type oprettet"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/afhentningstyper/{id}": {
            "get": {
                "tags": [
                    "Sager - Afhentnings Typer"
                ],
                "summary": "Find en specifik Afhentnings Type",
                "description": "",
                "operationId": "afhentningstypesSpecific",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Afhentnings Type ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "navn": {
                                            "type": "string",
                                            "example": "Bedste Afhentnings Type Navn"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 02:28:47"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "31-03-2022 03:15:46"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "Navn": "Bedste Afhentnings Type Navn",
                                        "created_at": "31-03-2022 02:28:47",
                                        "updated_at": "31-03-2022 03:15:46"
                                    }
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Ingen Afhentnings Type"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sager - Afhentnings Typer"
                ],
                "summary": "Opdater en Afhentnings Type",
                "description": "",
                "operationId": "afhentningstypesUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Afhentnings Type ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "navn",
                        "in": "query",
                        "description": "Afhentnings Type navnet",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Afhentnings Type Opdateret"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sager - Afhentnings Typer"
                ],
                "summary": "Slet en Afhentnings Type",
                "description": "",
                "operationId": "afhentningstypesDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Afhentnings Type ID'et",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Afhentnings Type Slettet"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}